another day, another grind missed. im slacking. day 5-1 finished

+++ AdventOfCode/Day5Map.java
@@ -11,7 +11,7 @@ public class Day5Map {
         ranges.add(range);
     }
 
-    public int give(int source){
+    public long give(long source){
         for(Day5Range range: ranges){
             if(range.contains(source)){
                 return range.give(source);
